+++ drives/item_items_item_workbook_worksheets_request_builder.go
@@ -29,9 +29,9 @@ type ItemItemsItemWorkbookWorksheetsRequestBuilderGetQueryParameters struct {
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int64 `uriparametername:"%24skip"`
+    Skip *int32 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // ItemItemsItemWorkbookWorksheetsRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemItemsItemWorkbookWorksheetsRequestBuilderGetRequestConfiguration struct {
