+++ groups/item_team_channels_get_all_retained_messages_request_builder.go
@@ -28,9 +28,9 @@ type ItemTeamChannelsGetAllRetainedMessagesRequestBuilderGetQueryParameters stru
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int64 `uriparametername:"%24skip"`
+    Skip *int32 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // ItemTeamChannelsGetAllRetainedMessagesRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemTeamChannelsGetAllRetainedMessagesRequestBuilderGetRequestConfiguration struct {
@@ -58,6 +58,9 @@ func NewItemTeamChannelsGetAllRetainedMessagesRequestBuilder(rawUrl string, requ
 // Deprecated: This method is obsolete. Use GetAsGetAllRetainedMessagesGetResponse instead.
 // returns a ItemTeamChannelsGetAllRetainedMessagesResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/channel-getallretainedmessages?view=graph-rest-1.0
 func (m *ItemTeamChannelsGetAllRetainedMessagesRequestBuilder) Get(ctx context.Context, requestConfiguration *ItemTeamChannelsGetAllRetainedMessagesRequestBuilderGetRequestConfiguration)(ItemTeamChannelsGetAllRetainedMessagesResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
@@ -78,6 +81,9 @@ func (m *ItemTeamChannelsGetAllRetainedMessagesRequestBuilder) Get(ctx context.C
 // GetAsGetAllRetainedMessagesGetResponse get all retained messages across all channels in a team. To learn more about how to use the Microsoft Teams export APIs to export content, see Export content with the Microsoft Teams export APIs.
 // returns a ItemTeamChannelsGetAllRetainedMessagesGetResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/channel-getallretainedmessages?view=graph-rest-1.0
 func (m *ItemTeamChannelsGetAllRetainedMessagesRequestBuilder) GetAsGetAllRetainedMessagesGetResponse(ctx context.Context, requestConfiguration *ItemTeamChannelsGetAllRetainedMessagesRequestBuilderGetRequestConfiguration)(ItemTeamChannelsGetAllRetainedMessagesGetResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
